Refactor: ExpenseServiceTest를 메서드별로 분리
